+++ scandeval/__init__.py
@@ -5,7 +5,7 @@ from .benchmark import Benchmark  # noqa
 from .datasets import load_dataset  # noqa
 from .utils import block_terminal_output
 
-__version__ = '1.3.8'  # noqa
+__version__ = '1.4.0'  # noqa
 
 # Block unwanted terminal outputs
 block_terminal_output()